+++ batoms/gui/gui_slicebatoms.py
@@ -15,7 +15,7 @@ class Batom_PT_prepare(Panel):
     bl_space_type = "VIEW_3D"
     bl_region_type = "UI"
     bl_options = {'DEFAULT_CLOSED'}
-    bl_category = "Species"
+    bl_category = "Batoms"
     bl_idname = "BATOMS_PT_Batom"
