+++ batoms/plugins/template/gui.py
@@ -49,7 +49,7 @@ class VIEW3D_PT_Batoms_template(Panel):
     bl_label = "template"
     bl_space_type = "VIEW_3D"
     bl_region_type = "UI"
-    bl_category = "Surface"
+    bl_category = "Plugins"
     bl_idname = "VIEW3D_PT_Batoms_template"
     bl_options = {'DEFAULT_CLOSED'}
